// src/app/features/contabilidad/domain/models/accounting-document.model.ts
export interface AccountingDocument {
  id: number;
  name: string;
  code: string;
  description: string;
  createdAt: string;
  updatedAt: string;
  deletedAt: string | null;
}

export interface AccountingDocumentPayload {
  name: string;
  code: string;
  description: string;
}

export interface AccountingDocumentPage {
  items: AccountingDocument[];
  links: {
    first: string | null;
    last: string | null;
    prev: string | null;
    next: string | null;
  };
  meta: {
    currentPage: number;
    from: number | null;
    lastPage: number;
    path: string;
    perPage: number;
    to: number | null;
    total: number;
  };
}